feat: Update AppBarSeparator to winui3/release/1.6.4
