+++ src/Uno.UI/UI/Xaml/Controls/AppBar/AppBarToggleButton.partial.mux.cs
@@ -27,12 +27,16 @@ namespace Microsoft.UI.Xaml.Controls;
 
 public partial class AppBarToggleButton : ToggleButton, ICommandBarElement, ICommandBarElement2, ICommandBarElement3, ICommandBarOverflowElement, ICommandBarLabeledElement, IAppBarButtonHelpersProvider
 {
-
 	DependencyProperty IAppBarButtonHelpersProvider.GetIsCompactDependencyProperty() => IsCompactProperty;
+
 	DependencyProperty IAppBarButtonHelpersProvider.GetUseOverflowStyleDependencyProperty() => UseOverflowStyleProperty;
+
 	DependencyProperty IAppBarButtonHelpersProvider.GetLabelPositionDependencyProperty() => LabelPositionProperty;
+
 	DependencyProperty IAppBarButtonHelpersProvider.GetLabelDependencyProperty() => LabelProperty;
+
 	DependencyProperty IAppBarButtonHelpersProvider.GetIconDependencyProperty() => IconProperty;
+
 	DependencyProperty IAppBarButtonHelpersProvider.GetKeyboardAcceleratorTextDependencyProperty() => KeyboardAcceleratorTextOverrideProperty;
 
 	/// <summary>
